requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
openai==1.61.1
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3